Creating import for function to reduce repeat code
